+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/trait/WithResult.java
@@ -8,5 +8,7 @@ public interface WithResult<T> {
     T read();
     Optional<T> readOpt();
     List<T> readAll();
+    long count();
+    boolean hasResults();
     WithResult<T> union(WithResult<T> union);
 }